+++ android/src/main/kotlin/com/plugin/piano_analytics_plugin/PAEvents.kt
@@ -1,0 +1,6 @@
+package com.plugin.piano_analytics_plugin
+
+object PAEvents {
+    const val SET_CONFIGURATION = "setConfiguration"
+    const val SEND_EVENT = "sendEvent"
+}