+++ lib/piano_analytics_plugin_method_channel.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'package:piano_analytics_plugin/piano_analytics_configuration_modes.dart';
 
 import 'piano_analytics_plugin_platform_interface.dart';
 
@@ -10,18 +11,48 @@ class MethodChannelPianoAnalyticsPlugin extends PianoAnalyticsPluginPlatform {
   final methodChannel = const MethodChannel('piano_analytics_plugin');
 
   /// Used to configure your app with piano analytics configuration.
+  /// Documentation : [https://developers.atinternet-solutions.com/piano-analytics/data-collection/sdks/android-java]
+  ///
   @override
   Future<void> setConfiguration({
     required String collectDomain,
     required int site,
-    String? privacyDefaultMode,
+    String? path,
+    String? customUserAgent,
+    bool crashDetection = true,
+    int sessionBackgroundDuration = 30,
+    bool ignoreLimitedAdvertisingTracking = false,
+    bool sendEventWhenOptout = true,
+    PrivacyDefaultMode privacyDefaultMode = PrivacyDefaultMode.optin,
+    OfflineEncryptionMode offlineEncryptionMode = OfflineEncryptionMode.force,
+    OfflineStorageMode offlineStorageMode = OfflineStorageMode.required,
+    int storageLifetimePrivacy = 395,
+    int storageLifetimeUser = 395,
+    int storageLifetimeVisitor = 395,
+    VisitorStorageMode visitorStorageMode = VisitorStorageMode.fixed,
+    VisitorIdTypeMode visitorIdTypeMode = VisitorIdTypeMode.uuid,
+    String? visitorId,
   }) async {
     await methodChannel.invokeMethod<void>(
       'setConfiguration',
       {
         'collectDomain': collectDomain,
         'site': site,
-        'privacyDefaultMode': privacyDefaultMode,
+        'path': path,
+        'customUserAgent': customUserAgent,
+        'crashDetection': crashDetection,
+        'sessionBackgroundDuration': sessionBackgroundDuration,
+        'ignoreLimitedAdvertisingTracking': ignoreLimitedAdvertisingTracking,
+        'sendEventWhenOptout': sendEventWhenOptout,
+        'privacyDefaultMode': privacyDefaultMode.name,
+        'offlineEncryptionMode': offlineEncryptionMode.name,
+        'offlineStorageMode': offlineStorageMode.name,
+        'storageLifetimePrivacy': storageLifetimePrivacy,
+        'storageLifetimeUser': storageLifetimeUser,
+        'storageLifetimeVisitor': storageLifetimeVisitor,
+        'visitorStorageMode': visitorStorageMode.name,
+        'visitorIdTypeMode': visitorIdTypeMode.name,
+        'visitorId': visitorId,
       },
     );
   }